fix: lazy-load discord carbon runtime for npm install
