+++ src/config/schema.base.generated.ts
@@ -27646,6 +27646,6 @@ export const GENERATED_BASE_CONFIG_SCHEMA: BaseConfigSchemaResponse = {
       tags: ["advanced", "url-secret"],
     },
   },
-  version: "2026.4.20-beta.1",
+  version: "2026.4.20-beta.2",
   generatedAt: "2026-03-22T21:17:33.302Z",
 };
